Centralize models in models.py
