+++ app.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, request, redirect, url_for, Response, flash, make_response
-from flask_sqlalchemy import SQLAlchemy
+from models import db, Cliente, Causa, FormatoLegal, Honorario, PagoCuota, Gasto
 import os
 import uuid
 from datetime import datetime, date
@@ -35,67 +35,7 @@ elif "?" not in DATABASE_URL and "sslmode=" not in DATABASE_URL:
 
 app.config["SQLALCHEMY_DATABASE_URI"] = DATABASE_URL
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-db = SQLAlchemy(app)
-
-# ========== MODELOS ==========
-class Cliente(db.Model):
-    __tablename__ = 'clientes'
-    id = db.Column(db.Integer, primary_key=True)
-    nombre = db.Column(db.String(100), nullable=False)
-    rut = db.Column(db.String(12), unique=True, nullable=False)
-    email = db.Column(db.String(100))
-    telefono = db.Column(db.String(20))
-    direccion = db.Column(db.String(200))
-    fecha_nacimiento = db.Column(db.Date)
-    causas = db.relationship('Causa', backref='cliente', lazy=True)
-    pagos = db.relationship('PagoCuota', backref='cliente', lazy=True)
-
-class Causa(db.Model):
-    __tablename__ = 'causas'
-    id = db.Column(db.Integer, primary_key=True)
-    tipo = db.Column(db.String(50))
-    rol = db.Column(db.String(20))
-    tribunal = db.Column(db.String(100))
-    cliente_id = db.Column(db.Integer, db.ForeignKey('clientes.id'), nullable=False)
-    formatos = db.relationship('FormatoLegal', backref='causa', lazy=True)
-
-class FormatoLegal(db.Model):
-    __tablename__ = 'formatos_legales'
-    id = db.Column(db.Integer, primary_key=True)
-    nombre_original = db.Column(db.String(255), nullable=False)
-    filename = db.Column(db.String(255), nullable=False, unique=True)
-    fecha_subida = db.Column(db.DateTime, default=datetime.utcnow)
-    usuario = db.Column(db.String(100))
-    causa_id = db.Column(db.Integer, db.ForeignKey('causas.id'))
-    version = db.Column(db.Integer, default=1)
-    observaciones = db.Column(db.String(255))
-
-class Honorario(db.Model):
-    __tablename__ = 'honorarios'
-    id = db.Column(db.Integer, primary_key=True)
-    cliente_id = db.Column(db.Integer, db.ForeignKey('clientes.id'), nullable=False)
-    causa_id = db.Column(db.Integer, db.ForeignKey('causas.id'), nullable=True)
-    descripcion = db.Column(db.String(255), nullable=False)
-    monto_total = db.Column(db.Float, nullable=False)
-    fecha_emision = db.Column(db.Date, default=datetime.utcnow)
-    en_cuotas = db.Column(db.Boolean, default=False)
-    numero_cuotas = db.Column(db.Integer, default=1)
-    estado = db.Column(db.String(50), default='pendiente')
-    pagos = db.relationship('PagoCuota', backref='honorario', lazy=True)
-
-class PagoCuota(db.Model):
-    __tablename__ = 'pagos_cuotas'
-    id = db.Column(db.Integer, primary_key=True)
-    honorario_id = db.Column(db.Integer, db.ForeignKey('honorarios.id'), nullable=False)
-    numero_cuota = db.Column(db.Integer)
-    monto_pagado = db.Column(db.Float)
-    fecha_pago = db.Column(db.Date)
-    vencimiento = db.Column(db.Date)
-    estado = db.Column(db.String(50), default='pendiente')
-    cliente_id = db.Column(db.Integer, db.ForeignKey('clientes.id'))
-
-with app.app_context():
-    db.create_all()
+db.init_app(app)
 
 # ========== FUNCIONES IA ==========
 def extraer_texto(path):
@@ -414,33 +354,6 @@ def exportar_facturacion():
 def servicio():
     return render_template("servicio.html")
 
-if __name__ == '__main__':
-    with app.app_context():
-        db.create_all()
-    app.run(debug=True)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-class Gasto(db.Model):
-    __tablename__ = 'gastos'
-    id = db.Column(db.Integer, primary_key=True)
-    descripcion = db.Column(db.String(255), nullable=False)
-    monto = db.Column(db.Float, nullable=False)
-    fecha = db.Column(db.Date, default=date.today)
-    categoria = db.Column(db.String(100))
-
 
 @app.route('/registrar_gasto', methods=['GET', 'POST'])
 def registrar_gasto():
@@ -458,17 +371,7 @@ def registrar_gasto():
 
     return render_template('registrar_gasto.html', date_today=date.today())
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+if __name__ == '__main__':
+    with app.app_context():
+        db.create_all()
+    app.run(debug=True)
